update: padding in the sidebar and the border on the terminal component
update: file structure for more seperation of concerns
