+++ src/ui/sidebar/chat-sidebar.html
@@ -27,7 +27,7 @@
           </button>
         </div>
 
-        <div class="px-2 py-2" *ngIf="!isCollapsed()">
+        <div class="px-3 py-2" *ngIf="!isCollapsed()">
           <div hlmInputGroup class="w-full">
             <div hlmInputGroupAddon>
               <ng-icon name="lucideSearch" />
